SBT: refactor tests, move those for cli and sysops
As a result, `tests` contains only `core` tests; the reason for it to
remain separate is due to special BuildInfo handling, to capture the
resource directories holding `.stat` files to test with.
