+++ scalafmt-cli/jvm/src/test/scala/org/scalafmt/cli/CliOptionsJVMTest.scala
