+++ scalafmt-cli/shared/src/test/scala/org/scalafmt/cli/FileTestOps.scala
